+++ src/main/java/lumi/insert/app/core/repository/projection/ProductRefund.java
@@ -1,6 +1,8 @@
 package lumi.insert.app.core.repository.projection;
 
+import java.math.BigDecimal;
+
 public record ProductRefund (
     String productName,
-    Long totalRefunded
+    BigDecimal totalRefunded
 ) {}
