+++ src/main/java/lumi/insert/app/core/repository/projection/ProductSale.java
@@ -1,6 +1,8 @@
 package lumi.insert.app.core.repository.projection;
 
+import java.math.BigDecimal;
+
 public record ProductSale (
     String productName,
-    Long totalSold
+    BigDecimal totalSold
 ) {}
