+++ src/main/java/lumi/insert/app/core/repository/projection/ProductStockProjection.java
@@ -1,5 +1,7 @@
 package lumi.insert.app.core.repository.projection;
 
+import java.math.BigDecimal;
+
 public interface ProductStockProjection {
-    Long getStockQuantity();
+    BigDecimal getStockQuantity();
 }
